==> 0-CleanDataset_wine.ipynb <==
!which python

import os
import json
import pickle
import pandas as pd
import numpy as np
from nltk.tokenize import sent_tokenize, word_tokenize, wordpunct_tokenize

# Load Dataset

df_data_file = "/u/pw7nc/experiments/Dataset/wine/df.pickle"
unpickled_df = pd.read_pickle(df_data_file)

unpickled_df

print("Number of users: {}".format(len(unpickled_df['userid'].unique())))
print("Number of items: {}".format(len(unpickled_df['itemid'].unique())))

# Check 0 sentence review

cnt = 0
for review_text in list(unpickled_df['review']):
    review_sents = sent_tokenize(review_text)
    if len(review_sents) == 0:
        try:
            assert review_text.strip() == ''
        except:
            print(review_text)
        cnt += 1
print("Number of reviews on whole dataset with 0 sentences: {}".format(cnt))

# Check Duplicate User-Item

user_item_set = set()
cnt_duplicate_review = 0

for idx, data_chunk in unpickled_df.iterrows():
    item_id = data_chunk['itemid']
    user_id = data_chunk['userid']
    if (user_id, item_id) in user_item_set:
        cnt_duplicate_review += 1
    else:
        user_item_set.add((user_id, item_id))

print("[Train] Number of duplicate reviews: {}".format(cnt_duplicate_review))
print("[Train] Number of unique reviews: {}".format(len(user_item_set)))

# Check Number of Reviews per User/Item

## User

groupby_user_whole = unpickled_df.groupby(['userid'])
groupby_item_whole = unpickled_df.groupby(['itemid'])

print("Number of unique user: {}".format(len(groupby_user_whole)))
print("Number of unique item: {}".format(len(groupby_item_whole)))

user_num_review_list = list()
user_num_review_dict = dict()
cnt = 0
for key, cur_user_df in groupby_user_whole:
    cur_user_id = key
    assert isinstance(cur_user_id, str)
    assert cur_user_id not in user_num_review_dict
    user_num_review_dict[cur_user_id] = len(cur_user_df)
    user_num_review_list.append(len(cur_user_df))
    cnt += 1
print("Finished! Totally {} users".format(cnt))

print("Mean number of review per user: {}".format(
    np.mean(user_num_review_list)
))
print("Min number of review per user: {}".format(
    np.min(user_num_review_list)
))
print("Max number of review per user: {}".format(
    np.max(user_num_review_list)
))

print("Top-10 least numbber of review per user: {}".format(
    sorted(user_num_review_list)[:10]
))
print("Top-10 most numbber of review per user: {}".format(
    sorted(user_num_review_list)[-10:]
))

## Item

item_num_review_list = list()
item_num_review_dict = dict()
cnt = 0
for key, cur_item_df in groupby_item_whole:
    cur_item_id = key
    assert isinstance(cur_item_id, str)
    assert cur_item_id not in item_num_review_dict
    item_num_review_dict[cur_item_id] = len(cur_item_df)
    item_num_review_list.append(len(cur_item_df))
    cnt += 1
print("Finished! Totally {} items".format(cnt))

print("Mean number of review per item: {}".format(
    np.mean(item_num_review_list)
))
print("Min number of review per item: {}".format(
    np.min(item_num_review_list)
))
print("Max number of review per item: {}".format(
    np.max(item_num_review_list)
))

print("Top-10 least numbber of review per item: {}".format(
    sorted(item_num_review_list)[:10]
))
print("Top-10 most numbber of review per item: {}".format(
    sorted(item_num_review_list)[-10:]
))

# Load Feature Words

feature_file_path = "/u/pw7nc/experiments/Dataset/wine/feature/feature_pure.txt"
feature_word_list = []
feature2id_dict = dict()
with open(feature_file_path, 'r') as f_feat:
    for line in f_feat:
        cur_feat_word = line.strip()
        assert cur_feat_word not in feature2id_dict
        # add this feature word into dict
        feature2id_dict[cur_feat_word] = str(len(feature_word_list))
        # add this feature word into list
        feature_word_list.append(cur_feat_word)
print("Totally {} features loaded.".format(len(feature2id_dict)))

# write feature words into file
feature2id_file_path = "/u/pw7nc/experiments/Dataset/wine/feature/feature2id.json"
with open(feature2id_file_path, 'w') as f_feat:
    json.dump(feature2id_dict, f_feat)

# Filter Dataset

def is_empty_review(review_text):
    """ Check whether a review is an empty review (i.e. 0 char or only have whitespaces)
    """
    review_sents = sent_tokenize(review_text)
    if len(review_sents) == 0:
        assert review_text.strip() == ''
        return True
    else:
        return False

def is_no_feature_review(review_text, feature_word_list):
    """ Check whether this review contains feature word
    """
    review_sents = sent_tokenize(review_text)
    feature_occurrence_flag = False
    for rvw_sent in review_sents:
        # tokenize this review sentence
        rvw_tokens = word_tokenize(rvw_sent.lower())
        for fea_word in feature_word_list:
            if fea_word in rvw_tokens:
                feature_occurrence_flag = True
                break
    return (not feature_occurrence_flag)

example_review = "this Wine is Dark. with Sweet Aroma."
print(is_no_feature_review(example_review, feature_word_list))

# 1. Remove Empty Review and No Feature Review
whole_review_data = []
cnt_lines = 0
cnt_num_empty_review = 0
cnt_num_no_feature_review = 0
for idx, data_chunk in unpickled_df.iterrows():
    item_id = data_chunk['itemid']
    user_id = data_chunk['userid']
    rating = data_chunk['rating']   # NOTE: rating's type is str
    review = data_chunk['review']
    if is_empty_review(review):
        # 1. check whether this review is empty
        cnt_num_empty_review += 1
    elif is_no_feature_review(review, feature_word_list):
        # 2. check whether this review contains
        cnt_num_no_feature_review += 1
    else:
        # add to list
        whole_review_data.append([item_id, user_id, rating, review])
    cnt_lines += 1
    if cnt_lines % 100000 == 0:
        print("{0} lines being processed. {1} empty reviews. {2} no feature reviews.".format(
            cnt_lines, cnt_num_empty_review, cnt_num_no_feature_review
        ))
print("{0} reviews has empty review.\n{1} reviews has no feature.\n{2} useful reviews.".format(
    cnt_num_empty_review, cnt_num_no_feature_review, len(whole_review_data)
))

# 2. Convert review to lower case and tokenize it.
print(whole_review_data[0])

# tokenize the lowercase the review text
# strip redundant whitespace of the rating (currently its still str since some of them is N/A)
for i in range(len(whole_review_data)):
    cur_origin_review = whole_review_data[i][-1]
    cur_tokenized_review = " ".join(word_tokenize(cur_origin_review.lower()))
    # remove redundant whitespace of itemid / userid / rating
    whole_review_data[i][0] = whole_review_data[i][0].strip()
    whole_review_data[i][1] = whole_review_data[i][1].strip()
    whole_review_data[i][2] = whole_review_data[i][2].strip()
    # tokenize the lowercase the review text
    whole_review_data[i][-1] = cur_tokenized_review
    if (i+1) % 200000 == 0:
        print("{} lines processed.".format(i+1))

print(whole_review_data[0])

print(len(whole_review_data))

# Check How many N/A ratings are there

cnt_na_rating = 0
for row in whole_review_data:
    if row[2] == "N/A":
        cnt_na_rating += 1
print("Totally {0} lines of data. Among them {1} lines has N/A rating.".format(
    len(whole_review_data), cnt_na_rating
))

# Get User/Item Statistics on the Filtered Dataset

df_filtered_review = pd.DataFrame(whole_review_data, columns=['item', 'user', 'rating', 'review'])

df_filtered_review.head()

groupby_user_whole_filter = df_filtered_review.groupby(['user'])
groupby_item_whole_filter = df_filtered_review.groupby(['item'])

print("Number of unique user: {}".format(len(groupby_user_whole_filter)))
print("Number of unique item: {}".format(len(groupby_item_whole_filter)))

## User

user_num_review_filter_list = list()
user_num_review_filter_dict = dict()
cnt = 0
for key, cur_user_df in groupby_user_whole_filter:
    cur_user_id = key
    assert isinstance(cur_user_id, str)
    assert cur_user_id not in user_num_review_filter_list
    user_num_review_filter_dict[cur_user_id] = len(cur_user_df)
    user_num_review_filter_list.append(len(cur_user_df))
    cnt += 1
print("Finished! Totally {} users".format(cnt))

print("Mean number of review per user: {}".format(
    np.mean(user_num_review_filter_list)
))
print("Min number of review per user: {}".format(
    np.min(user_num_review_filter_list)
))
print("Max number of review per user: {}".format(
    np.max(user_num_review_filter_list)
))

print("Top-10 least numbber of review per user: {}".format(
    sorted(user_num_review_filter_list)[:10]
))
print("Top-10 most numbber of review per user: {}".format(
    sorted(user_num_review_filter_list)[-10:]
))

## Item

item_num_review_filter_list = list()
item_num_review_filter_dict = dict()
cnt = 0
for key, cur_item_df in groupby_item_whole_filter:
    cur_item_id = key
    assert isinstance(cur_item_id, str)
    assert cur_item_id not in item_num_review_filter_dict
    item_num_review_filter_dict[cur_item_id] = len(cur_item_df)
    item_num_review_filter_list.append(len(cur_item_df))
    cnt += 1
print("Finished! Totally {} items".format(cnt))

print("Mean number of review per item: {}".format(
    np.mean(item_num_review_filter_list)
))
print("Min number of review per item: {}".format(
    np.min(item_num_review_filter_list)
))
print("Max number of review per item: {}".format(
    np.max(item_num_review_filter_list)
))

print("Top-10 least numbber of review per item: {}".format(
    sorted(item_num_review_filter_list)[:10]
))
print("Top-10 most numbber of review per item: {}".format(
    sorted(item_num_review_filter_list)[-10:]
))

# Save Cleaned Dataset into File

whole_data_filtered_file = "/u/pw7nc/experiments/Dataset/wine/whole_cleaned.json"
with open(whole_data_filtered_file, 'w') as f_out:
    for row in whole_review_data:
        row_dict = {
            'user': row[1],
            'item': row[0],
            'rating': row[2],
            'review': row[3]
        }
        # dump this dict into file
        json.dump(row_dict, f_out)
        f_out.write('\n')
