==> 2-SplitDataset_wine.ipynb <==
!which python

import nltk
import spacy
import re
import json
import pandas as pd
import os
import numpy as np
from googletrans import Translator
from nltk.tokenize import sent_tokenize, word_tokenize, wordpunct_tokenize

# Read Data

# Load the whole filtered data
dir_path = '../Dataset/wine/'
# Load whole dataset (cleaned and filtered)
whole_review = []
cnt = 0
file_path = os.path.join(dir_path, 'whole_filtered.json')
with open(file_path) as f:
    for line in f:
        line_data = json.loads(line)
        user_id = line_data['user']
        item_id = line_data['item']
        rating = line_data['rating']
        review = line_data['review']
        whole_review.append([item_id, user_id, rating, review])
        cnt += 1
        if cnt % 100000 == 0:
            print("{} lines loaded.".format(cnt))
print("Finish loading whole cleaned filtered dataset, totally {} lines.".format(cnt))

# detector = Translator()
# cnt_none_english = 0
# none_english_idx = list()
# none_english_lans = list()

# for idx, rvw_data in enumerate(whole_review):
#     rvw_text = rvw_data[-1]
#     dec_lan = detector.detect(rvw_text)
#     if dec_lan.lang == "en":
#         pass
#     else:
#         cnt_none_english += 1
#         none_english_idx.append(idx)
#         none_english_lans.append(dec_lan.lang)
#     if (idx+1) % 10000 == 0:
#         print("{} lines".format(idx+1))


# Split the Data into Train/Test

user_group = dict()
for rvw_data in whole_review:
    user_id = rvw_data[1]
    if user_id not in user_group:
        user_group[user_id] = list()
    user_group[user_id].append(rvw_data)

len(user_group)

user_group = {u: v for u, v in user_group.items()}

TRAIN_FILE = os.path.join(dir_path, 'train_filtered.json')
VALID_FILE = os.path.join(dir_path, 'valid_filtered.json')
TEST_FILE = os.path.join(dir_path, 'test_filtered.json')

import random
import math
from collections import Counter

VALID_RATIO = 0.00
TEST_RATIO = 0.20

train_lines, valid_lines, test_lines = [], [], []

train_item_counter = Counter()
valid_item_counter = Counter()
test_item_counter = Counter()

cnt_user_num = 0

for uid, lines in user_group.items():
    random.shuffle(lines)

    leng = len(lines)
    valid_num = math.floor(leng * VALID_RATIO)
    test_num = math.floor(leng * TEST_RATIO)

    valid_lines += lines[:valid_num]
    test_lines += lines[-test_num:]
    train_lines += lines[valid_num:-test_num]

    for rvw_data in lines[:valid_num]:
        iid = rvw_data[0]
        valid_item_counter.update([iid])
    for rvw_data in lines[-test_num:]:
        iid = rvw_data[0]
        test_item_counter.update([iid])
    for rvw_data in lines[valid_num:-test_num]:
        iid = rvw_data[0]
        train_item_counter.update([iid])

    cnt_user_num += 1
    if cnt_user_num % 1000 == 0:
        print("{} users".format(cnt_user_num))

print("Number of review instance in train: {0}\nNumber of review instance in valid: {1}\nNumber of review instance in test: {2}".format(len(train_lines), len(valid_lines), len(test_lines)))

train_item_set = set(train_item_counter.keys())
valid_item_set = set(valid_item_counter.keys())
test_item_set = set(test_item_counter.keys())

for iid in test_item_set:
    if iid not in train_item_set:
        print(iid)
    else:
        if train_item_counter[iid] <= test_item_counter[iid]:
            print("iid: {}\tnum on train: {}\tnum on test: {}".format(
                iid, train_item_counter[iid], test_item_counter[iid]
            ))

with open(TRAIN_FILE, 'w') as train_file, open(VALID_FILE, 'w') as valid_file, open(TEST_FILE, 'w') as test_file:
    for rvw_data in train_lines:
        row_iid = rvw_data[0]
        row_uid = rvw_data[1]
        row_rating = rvw_data[2]
        row_review = rvw_data[3]
        row_dict = {
            'user': row_uid,
            'item': row_iid,
            'rating': row_rating,
            'review': row_review
        }
        # dump this into file
        json.dump(row_dict, train_file)
        train_file.write("\n")
    for rvw_data in test_lines:
        row_iid = rvw_data[0]
        row_uid = rvw_data[1]
        row_rating = rvw_data[2]
        row_review = rvw_data[3]
        row_dict = {
            'user': row_uid,
            'item': row_iid,
            'rating': row_rating,
            'review': row_review
        }
        # dump this into file
        json.dump(row_dict, test_file)
        test_file.write("\n")

train_lines[0]
