==> 3-SentenceUserItem2Feature_ta.ipynb <==
!which python

import re
import json
import os

import numpy as np
import pandas as pd
from sklearn.preprocessing import normalize
from sklearn.feature_extraction.text import CountVectorizer,TfidfTransformer

import spacy
import nltk
from nltk.tokenize import sent_tokenize, word_tokenize
from spacy.lang.en import English
nlp = English()
# Create a Tokenizer with the default settings for English
# including punctuation rules and exceptions
tokenizer = nlp.tokenizer
import string
punct = string.punctuation
from sklearn.feature_extraction import _stop_words

dataset_name = "tripadvisor"

punct

# Read Data

dir_path = '../Dataset/{}'.format(dataset_name)
# Load train dataset
train_review = []
cnt = 0
file_path = os.path.join(dir_path, 'train_review_filtered.json')
with open(file_path) as f:
    print("Load file: {}".format(file_path))
    for line in f:
        line_data = json.loads(line)
        user_id = line_data['user']
        item_id = line_data['item']
        rating = line_data['rating']
        review = line_data['review']
        train_review.append([item_id, user_id, rating, review])
        cnt += 1
        if cnt % 50000 == 0:
            print('{} lines loaded.'.format(cnt))
print('Finish loading train dataset, totally {} lines.'.format(len(train_review)))
# Load test dataset
test_review = []
cnt = 0
file_path = os.path.join(dir_path, 'test_review_filtered_clean.json')
with open(file_path) as f:
    print("Load file: {}".format(file_path))
    for line in f:
        line_data = json.loads(line)
        user_id = line_data['user']
        item_id = line_data['item']
        rating = line_data['rating']
        review = line_data['review']
        test_review.append([item_id, user_id, rating, review])
        cnt += 1
        if cnt % 10000 == 0:
            print('{} lines loaded.'.format(cnt))
print('Finish loading test dataset, totally {} lines.'.format(len(test_review)))

## Convert List Data to Pandas Dataframe

df_train_data = pd.DataFrame(train_review, columns=['item', 'user', 'rating', 'review'])
df_test_data = pd.DataFrame(test_review, columns=['item', 'user', 'rating', 'review'])

df_train_data

print("Number of users on train: {}\tNumber of items on train: {}".format(
    len(df_train_data['user'].unique()), len(df_train_data['item'].unique())
))
print("Number of users on test: {}\tNumber of items on test: {}".format(
    len(df_test_data['user'].unique()), len(df_test_data['item'].unique())
))

# Compute Sentence Tf-idf

def catDoc(textlist):
    res = []
    for tlist in textlist:
        res.extend(tlist)
    return res

def get_tfidf_embedding(text, feature_word_list):
    """
    :param: text: list, sent_number * word
    :return: 
        vectorizer: 
            vocabulary_: word2id
            get_feature_names(): id2word
        tfidf: array [sent_number, max_word_number]
    """
    vectorizer = CountVectorizer(lowercase=True, vocabulary=feature_word_list)
    word_count = vectorizer.fit_transform(text)
    tfidf_transformer = TfidfTransformer()
    tfidf = tfidf_transformer.fit_transform(word_count)
    tfidf_weight = tfidf.toarray()
    return vectorizer, tfidf_weight

def get_tf_score(text, feature_word_list):
    vectorizer = CountVectorizer(lowercase=True, vocabulary=feature_word_list)
    word_count = vectorizer.fit_transform(text)
    return word_count.toarray()

def get_df_score(text, feature_word_list):
    vectorizer = CountVectorizer(lowercase=True, vocabulary=feature_word_list)
    word_count = vectorizer.fit_transform(text)
    # from word count (i.e. tf) get document frequency (i.e. df)
    df_count = np.sum(word_count.toarray()>0, axis=0)
    return df_count

def compress_array(a, id2word, vocab):
    """
    :param a: matrix, [N, M], N is document number, M is word number
    :param id2word: word id to word
    :return: 
    """
    d = {}
    # Loop over documents
    for i in range(len(a)):
        d[i] = {}
        # Loop over words
        for j in range(len(a[i])):
            if a[i][j] != 0:
                wid_voc = vocab[id2word[j]]
                d[i][wid_voc] = a[i][j]
    return d

# Load Feature Words

feature_2_id_file = '../Dataset/{}/train/feature/feature2id.json'.format(dataset_name)
with open(feature_2_id_file, 'r') as f:
    print("Load file: {}".format(feature_2_id_file))
    feature_vocab = json.load(f)

len(feature_vocab)

feature_vocab['wifi']

feature_word_list = list(feature_vocab.keys())
print('Number of feature words: {}'.format(len(feature_word_list)))

id2feature_dict = dict()
for key,value in feature_vocab.items():
    id2feature_dict[value] = key

id_2_feature_file = '../Dataset/{}/train/feature/id2feature.json'.format(dataset_name)
with open(id_2_feature_file, 'w') as f:
    print("Write file: {}".format(id_2_feature_file))
    json.dump(id2feature_dict, f)

# Check Whether there are reviews with no sentences

invalid_data = 0
for idx, row in df_train_data.iterrows():
    review_text = row['review']
    review_sents = sent_tokenize(review_text)
    if len(review_sents) == 0:
        print(row)
        invalid_data += 1
print("Invalid reviews: {}".format(invalid_data))

# Construct Sentence Vocab

# sentence vocab
sentence_count = dict()
sentence_with_no_feature = 0
# Loop for each review
for idx, row in df_train_data.iterrows():
    review_text = row['review']
    review_sents = sent_tokenize(review_text)
    tf_score = get_tf_score(review_sents, feature_word_list)
    # _, tf_score = get_tfidf_embedding(review_sents, feature_word_list)
    # Sum up the tf-value for each sentence so that if this sum is 0, this sentence should be removed
    tfidf_sum_sents = np.sum(tf_score, axis=1)
    for i in range(len(review_sents)):
        if tfidf_sum_sents[i] != 0.0:
            cur_sent = review_sents[i]
            # check whether this sentence has more than 3 tokens
            tokens = word_tokenize(cur_sent)
            cnt_tokens = 0
            for token in tokens:
                if token.isdigit() or (token in punct):
                    pass
                else:
                    cnt_tokens += 1
            # only sentence with more than (or equal to) 2 effective tokens 
            # can be added into the sentence vocab
            if cnt_tokens < 2:
                pass
            else:
                sentence_count[cur_sent] = 1 + sentence_count.get(cur_sent, 0)
        else:
            sentence_with_no_feature += 1
    if (idx+1) % 10000 == 0:
        print("Processed {} lines".format(idx+1))
print('Finish. Totally {} lines'.format(idx+1))

print("Number of sentences with feature word(s): {0}\nNumber of sentences w/o feature word: {1}".format(
    len(sentence_count), sentence_with_no_feature
))

# sort sentence based on counts (the majority should be 1)
sorted_sent_counts = sorted(sentence_count.items(), key = lambda x: -x[1])

# sentence_vocab_list = list(sentence_count.keys())
# Building mappings from sentences to ids and ids to sentences
sent_to_id = {entry[0]: str(id) for (id, entry) in enumerate(sorted_sent_counts)}
# Since we loaded all the tokenized sentences, we don't need to add the special UNK token
id_to_sent = {str(id): sent for (sent, id) in sent_to_id.items()}

assert len(sent_to_id) == len(id_to_sent)

id_to_sent['0']

id2sentence_filepath = '../Dataset/{}/train/sentence/id2sentence.json'.format(dataset_name)
with open(id2sentence_filepath, 'w') as f:
    print("Write file: {}".format(id2sentence_filepath))
    json.dump(id_to_sent, f)

sentence2id_filepath = '../Dataset/{}/train/sentence/sentence2id.json'.format(dataset_name)
with open(sentence2id_filepath, 'w') as f:
    print("Write file: {}".format(sentence2id_filepath))
    json.dump(sent_to_id, f)

# Load id2sentence and sentence2id, check whether they are the same as the newly processed mappings
id2sentence_filepath = '../Dataset/{}/train/sentence/id2sentence.json'.format(dataset_name)
with open(id2sentence_filepath, 'r') as f:
    print("Load file: {}".format(id2sentence_filepath))
    trainset_id2sent = json.load(f)
sentence2id_filepath = '../Dataset/{}/train/sentence/sentence2id.json'.format(dataset_name)
with open(sentence2id_filepath, 'r') as f:
    print("Load file: {}".format(sentence2id_filepath))
    trainset_sent2id = json.load(f)

assert trainset_id2sent == id_to_sent
assert trainset_sent2id == sent_to_id

# Get Sentence Feature

df_train_data

def check_vocab_is_same(sklearn_vocab, feature_vocab):
    if len(sklearn_vocab) == len(feature_vocab):
        for key, value in sklearn_vocab.items():
            sklearn_vocab_id = value
            feature_vocab_id = feature_vocab[key]
            if int(feature_vocab_id) == sklearn_vocab_id:
                continue
            else:
                return False
    else:
        return False
    return True

sentence_text_list = list(sent_to_id.keys())

len(sentence_text_list)

sentence_text_list[:10]

cntvector, tfidf_weight = get_tfidf_embedding(sentence_text_list, feature_word_list)

df_count = get_df_score(sentence_text_list, feature_word_list)

df_count.shape

trainset_feature_df = dict()
trainset_feature_df_norm = dict()
for i in range(len(feature_word_list)):
    trainset_feature_df[feature_word_list[i]] = df_count[i]
    trainset_feature_df_norm[feature_word_list[i]] = df_count[i]/len(sentence_text_list)

type(trainset_feature_df['wifi'])

for key, value in trainset_feature_df.items():
    if isinstance(value, np.int64):
        trainset_feature_df[key] = int(value)

type(trainset_feature_df['wifi'])

trainset_feat_df_file = '../Dataset/{}/train/feature/feature2df.json'.format(dataset_name)

with open(trainset_feat_df_file, 'w') as f:
    print("Write file: {}".format(trainset_feat_df_file))
    json.dump(trainset_feature_df, f)

trainset_feature_df_sort = dict(sorted(trainset_feature_df.items(), key = lambda x: -x[1]))

trainset_feature_df_sort_list = list(trainset_feature_df_sort.keys())
trainset_feature_df_sort_rank = dict()
for i in range(len(trainset_feature_df_sort_list)):
    trainset_feature_df_sort_rank[trainset_feature_df_sort_list[i]] = i+1

this_word = 'wifi'
print("df value: {}".format(trainset_feature_df[this_word]))
print("rank of the feature: {}".format(trainset_feature_df_sort_rank[this_word]))

tfidf_weight.shape

check_vocab_is_same(cntvector.vocabulary_, feature_vocab)

sentence_to_feature = dict()
sentence_with_no_feature = 0
tfidf_sum_sents = np.sum(tfidf_weight, axis=1)
print("Shape of tf-idf sum: {}".format(tfidf_sum_sents.shape))
for i in range(len(sentence_text_list)):
    cur_sent = sentence_text_list[i]
    # if this sentence is in the sent_to_id vocabulary
    assert cur_sent in sent_to_id
    # get the sentence_id (str)
    cur_sent_id = sent_to_id[cur_sent]
    assert int(cur_sent_id) == i
    # find all the feature that has non-zero tf-idf weight
    feature_dict = dict()
    for j in range(len(tfidf_weight[i])):
        if tfidf_weight[i][j] != 0.0:
            # get the feature
            feature_id = str(j)
            feature = feature_word_list[j]
            feature_tfidf = tfidf_weight[i][j]
            feature_dict[feature_id] = feature_tfidf
    if len(feature_dict) > 0:
        sentence_to_feature[cur_sent_id] = feature_dict
    else:
        sentence_with_no_feature += 1
    if (i+1) % 50000 == 0:
        print("Processed {} lines".format(i+1))
print("Finish. Totally {} lines".format(i+1))
print("Totally {} sentences has at least 1 feature and {} sentences don't have feature.".format(
    len(sentence_to_feature), sentence_with_no_feature))

sentence2feature_filepath = '../Dataset/{}/train/sentence/sentence2feature.json'.format(dataset_name)
with open(sentence2feature_filepath, 'w') as f:
    print("Write file: {}".format(sentence2feature_filepath))
    json.dump(sentence_to_feature, f)

sentence_to_feature['0']

id_to_sent['0']

id2feature_dict['9']

num_feature_per_sentence = []
for key, value in sentence_to_feature.items():
    num_feature_per_sentence.append(len(value))
    assert len(value) > 0

print("Mean number of features per sentence: {}".format(np.mean(num_feature_per_sentence)))
print("Max number of features per sentence: {}".format(np.max(num_feature_per_sentence)))
print("Min number of features per sentence: {}".format(np.min(num_feature_per_sentence)))

# Get User to Feature

df_train_data

## GroupBy User

group_by_user = df_train_data.groupby('user')
user_id_list = []
user_reviews = []
# Loop over all user
for user_df_chunk in list(group_by_user):
    user_id = int(user_df_chunk[0])
    user_df = user_df_chunk[1]
    user_text = " ".join(list(user_df['review']))
    user_id_list.append(user_id)
    user_reviews.append(user_text)

print("Number of users: {}".format(len(user_id_list)))

assert len(user_id_list) == len(user_reviews)

## Compute User Tf-idf

cntvector_user, tfidf_weight_user = get_tfidf_embedding(user_reviews, feature_word_list)

check_vocab_is_same(cntvector_user.vocabulary_, feature_vocab)

tfidf_weight_user.shape

print(feature_word_list[:20])

user_to_feature = dict()
for i in range(len(user_id_list)):
    feature_dict = dict()
    cur_user_id = user_id_list[i]
    assert len(tfidf_weight_user[i]) == len(feature_vocab)
    for j in range(len(tfidf_weight_user[i])):
        if tfidf_weight_user[i][j] != 0.0:
            # get the feature
            # NOTE: make sure that the feature_id is str format
            feature_id = str(j)
            feature = feature_word_list[j]
            assert feature_vocab[feature] == feature_id
            feature_tfidf = tfidf_weight_user[i][j]
            feature_dict[feature_id] = feature_tfidf
    assert len(feature_dict) > 0
    user_to_feature[str(cur_user_id)] = feature_dict
    if (i+1) % 1000 == 0:
        print("{} user processed.".format(i+1))
print("Totally {} users".format(i+1))

len(user_to_feature)

num_feature_per_user = []
for key,value in user_to_feature.items():
    num_feature_per_user.append(len(value))
    assert len(value) > 0

print("Mean number of features per user: {}".format(np.mean(num_feature_per_user)))
print("Max number of features per user: {}".format(np.max(num_feature_per_user)))
print("Min number of features per user: {}".format(np.min(num_feature_per_user)))

len(user_to_feature['1001'])

## Save User to Feature Mapping into Json File

user2feature_filepath = '../Dataset/{}/train/user/user2feature.json'.format(dataset_name)
with open(user2feature_filepath, 'w') as f:
    print("Write file: {}".format(user2feature_filepath))
    json.dump(user_to_feature, f)

# Get Item to Feature

## GroupBy Item

group_by_item = df_train_data.groupby('item')
item_id_list = []
item_reviews = []
# Loop over all user
for item_df_chunk in list(group_by_item):
    item_id = str(item_df_chunk[0])
    item_df = item_df_chunk[1]
    item_text = " ".join(list(item_df['review']))
    item_id_list.append(item_id)
    item_reviews.append(item_text)

print("Number of items: {}".format(len(item_id_list)))

assert len(item_id_list) == len(item_reviews)

## Compute Item Tf-idf

cntvector_item, tfidf_weight_item = get_tfidf_embedding(item_reviews, feature_word_list)

check_vocab_is_same(cntvector_item.vocabulary_, feature_vocab)

tfidf_weight_item.shape

item_to_feature = dict()
for i in range(len(item_id_list)):
    feature_dict = dict()
    cur_item_id = item_id_list[i]
    assert len(tfidf_weight_item[i]) == len(feature_vocab)
    for j in range(len(tfidf_weight_item[i])):
        if tfidf_weight_item[i][j] != 0.0:
            # get the feature
            feature_id = str(j)
            feature = feature_word_list[j]
            assert feature_id == feature_vocab[feature]
            feature_tfidf = tfidf_weight_item[i][j]
            feature_dict[feature_id] = feature_tfidf
    assert len(feature_dict) > 0
    item_to_feature[cur_item_id] = feature_dict
    if (i+1) % 1000 == 0:
        print("{} items processed.".format(i+1))
print('Finish. Totally {} items'.format(i+1))

len(item_to_feature)

num_feature_per_item = []
for key,value in item_to_feature.items():
    num_feature_per_item.append(len(value))
    assert len(value) > 0

print("Mean number of features per item: {}".format(np.mean(num_feature_per_item)))
print("Max number of features per item: {}".format(np.max(num_feature_per_item)))
print("Min number of features per item: {}".format(np.min(num_feature_per_item)))

item2feature_filepath = '../Dataset/{}/train/item/item2feature.json'.format(dataset_name)
with open(item2feature_filepath, 'w') as f:
    print("Write file: {}".format(item2feature_filepath))
    json.dump(item_to_feature, f)

# Compute Top User/Item Features

# TODO: Sanity Check
