==> 5-FinetuneGloveFeatureWords_yelp.ipynb <==
!which python

import csv
import numpy as np
from collections import Counter
from nltk.corpus import brown
from mittens import GloVe, Mittens
from sklearn.feature_extraction import _stop_words
from sklearn.feature_extraction.text import CountVectorizer
import pickle
import re
import json
import os
import pandas as pd
import spacy
import nltk
from nltk.tokenize import sent_tokenize, word_tokenize

def glove2dict(glove_filename):
    with open(glove_filename, encoding='utf-8') as f:
        reader = csv.reader(f, delimiter=' ', quoting=csv.QUOTE_NONE)
        embed = {line[0]: np.array(list(map(float, line[1:])))
                for line in reader}
    return embed

def get_rareoov(xdict, val):
    return [k for (k,v) in Counter(xdict).items() if v<=val]

dataset_name = "yelp"

# Load Data

# Load the orignial dataset to train glove. Larger dataset can give better results for word embedding
# NOTE: For current fast development use the train_review_filtered dataset.
# Later should change this to whole dataset.

dir_path = '../Dataset/{}/'.format(dataset_name)
# Load train dataset
train_review = []
cnt = 0
file_path = os.path.join(dir_path, 'train_review_filtered.json')
with open(file_path) as f:
    print("Load file: {}".format(file_path))
    for line in f:
        line_data = json.loads(line)
        user_id = line_data['user']
        item_id = line_data['item']
        rating = line_data['rating']
        review = line_data['review']
        train_review.append([item_id, user_id, rating, review])
        cnt += 1
        if cnt % 100000 == 0:
            print('{} lines loaded.'.format(cnt))
print('Finish loading train dataset, totally {} lines.'.format(len(train_review)))

df_train_data = pd.DataFrame(train_review, columns=['item', 'user', 'rating', 'review'])

train_review_doc = list(df_train_data['review'])
assert len(train_review_doc) == len(df_train_data)

with open('./embeddings/yelp/train-review.txt', 'w') as f:
    cnt = 0
    for review in train_review_doc:
        # write this into file
        f.write(review.strip())
        f.write('\n')
        cnt += 1
    print('Write {} lines of review into file'.format(cnt))

from collections import Counter
import simplejson
from nltk.tokenize import word_tokenize
import string
import re
import nltk
nltk.download('punkt')

def tokenize(infname="embeddings/yelp/train-review.txt"):
    outfname = open(infname + ".tokenized", "w")
    sents = open(infname).read().split("\n")
    for sent in sents:
        # convert characters into lower cases
        text = sent.lower()
        # tokenize the raw text using nltk.tokenize
        text = word_tokenize(text)
        # remove all punctuation
        text = list(filter(lambda token: token not in string.punctuation, text))
        # # replace all numbers with a special token <num>
        # text = [re.sub('[0-9]+','<num>', token) for token in text]
        # write this processed text into the output file
        simplejson.dump(text, outfname)
        outfname.write("\n")
    outfname.close()


def token_filter(infname="embeddings/yelp/train-review.txt.tokenized", thresh=5):
    outfname = open(infname.replace(".tokenized", ".filtered"), 'w')
    vocab = []  
    vocab_cnt = Counter()
    tokenized_sents = []
    with open(infname) as f_in:
        for line in f_in:
            tokenized_text = simplejson.loads(line)
            vocab_cnt.update(tokenized_text)
            tokenized_sents.append(tokenized_text)
    for tokenized_sent in tokenized_sents:
        # filter word that has less than 5 TF over the whole dataset
        filtered_text = []
        for token in tokenized_sent:
            if vocab_cnt[token] < thresh:
                pass
            else:
                filtered_text.append(token)
        # write this filtered data into file
        # ignore the last line of the initial imdb-small dataset since it's an empty line
        if len(tokenized_sent) > 0:
            new_sentence = " ".join(filtered_text).strip()+'\n'
            outfname.write(new_sentence)
    # bulld the vocabulary
    ## filter out token with low TF
    vocab_cnt_filter = {token: cnt for token, cnt in vocab_cnt.items() if cnt >= thresh}
    vocab_cnt_filter = Counter(vocab_cnt_filter)
    ## convert this vocabulary counter to a list
    for token,cnt in vocab_cnt_filter.most_common():
        vocab.append(token)
    outfname.close()
    return vocab

tokenize()
print("Finish tokenize ...")
vocab = token_filter(thresh=10)
print("Finish construct vocab !")
print("The vocab size = {}".format(len(vocab)))

## Load Feature Words

feature_2_id_file = '../Dataset/{}/train/feature/feature2id.json'.format(dataset_name)
with open(feature_2_id_file, 'r') as f:
    print("Load file: {}".format(feature_2_id_file))
    feature_vocab = json.load(f)
print("Size of the feature vocab: {}".format(len(feature_vocab)))

for fea_word in list(feature_vocab.keys()):
    if fea_word in vocab:
        pass
    else:
        print(fea_word)

total_vocab = set(vocab) | set(list(feature_vocab.keys()))
print("The total vocab size = {}".format(len(total_vocab)))

total_vocab_list = list(total_vocab)

cv = CountVectorizer(ngram_range=(1,1), vocabulary=total_vocab_list)
X = cv.fit_transform(train_review_doc)
Xc = (X.T * X)
Xc.setdiag(0)
coocc_ar = Xc.toarray()

coocc_ar.shape

X.shape

len(cv.vocabulary_)

train_vocab = cv.get_feature_names()

print(train_vocab[:20])

cv.vocabulary_["wifi"]

train_vocab == total_vocab_list

from mittens import GloVe
import numpy as np

# mittens_model = Mittens(n=256, max_iter=1000)

# new_embeddings = mittens_model.fit(
#     coocc_ar,
#     vocab=train_vocab)

# NOTE: For fast development, set the max_iter to be 100.
# Should set it to be 1000 for more accurate word embedding.
glove_model = GloVe(n=256, max_iter=400)
embeddings = glove_model.fit(coocc_ar)

embeddings.shape

len(cv.vocabulary_)

len(cv.get_feature_names())

vocabulary_words = cv.get_feature_names()

print(list(feature_vocab.keys())[:20])

print(vocabulary_words[:20])

# construct an embedding dict which has a format of {'word':'wine', 'embedding': array([...])}
word_embedding_dict = dict()
for i in range(len(embeddings)):
    current_word = vocabulary_words[i]
    current_embedding = embeddings[i]
    word_embedding_dict[current_word] = current_embedding
print("Totally {} words begin added in the word embedding dict".format(len(word_embedding_dict)))

# Save the Embedding dict into pickle file
import pickle
with open('./embeddings/yelp/glove.pickle', 'wb') as handle:
    pickle.dump(word_embedding_dict, handle, protocol=pickle.HIGHEST_PROTOCOL)

# extract the embedding vector for each feature word
feature_embedding = dict()
emb_cnt = 0
for feature_word in list(feature_vocab.keys()):
    feature_word_id = cv.vocabulary_[feature_word]
    feature_vocab_id = feature_vocab[feature_word]
    feature_word_emb = embeddings[feature_word_id]
    feature_embedding[feature_vocab_id] = feature_word_emb
    emb_cnt += 1
print("Totally {} feature words".format(emb_cnt))

# extract the embedding vector for each feature word
feature_embed_dict = dict()
for emb_chunk in list(feature_embedding.items()):
    feature_id = emb_chunk[0]   # this is a str
    feature_emb_np = emb_chunk[1].tolist()
    feature_embed_dict[feature_id] = feature_emb_np

len(feature_embed_dict)

feature_emb_filepath = '../Dataset/yelp/train/feature/featureid2embedding.json'
with open(feature_emb_filepath, 'w') as f:
    print("Write file: {}".format(feature_emb_filepath))
    json.dump(feature_embed_dict, f)

# Transfer Previously Trained Glove Embedding
## Load Previous Feature2ID and ID2Feature Mapping

# Feature2id and id2feature
feat2id_file = '../Dataset/yelp_large/train/feature/feature2id.json'.format(dataset_name)
with open(feat2id_file, 'r') as f:
    print("Load file: {}".format(feat2id_file))
    feature_to_id = json.load(f)
id2feat_file = '../Dataset/yelp_large/train/feature/id2feature.json'.format(dataset_name)
with open(id2feat_file, 'r') as f:
    print("Load file: {}".format(id2feat_file))
    id_to_feature = json.load(f)
print("Number of features: {}".format(len(id_to_feature)))

## Load Previous Featureid2Embedding Mapping

featid2emb_file = '../Dataset/yelp_large/train/feature/featureid2embedding.json'.format(dataset_name)
with open(featid2emb_file, 'r') as f:
    print("Load file: {}".format(featid2emb_file))
    featureid2emb_voc = json.load(f)

# Feature2id and id2feature
feat2id_file = '../Dataset/{}/train/feature/feature2id.json'.format(dataset_name)
with open(feat2id_file, 'r') as f:
    print("Load file: {}".format(feat2id_file))
    feat2id_voc = json.load(f)
id2feat_file = '../Dataset/{}/train/feature/id2feature.json'.format(dataset_name)
with open(id2feat_file, 'r') as f:
    print("Load file: {}".format(id2feat_file))
    id2feat_voc = json.load(f)
print("Number of features: {}".format(len(id2feat_voc)))

featid2emb_voc = dict()
tracked_prev_feat_ids = set()
for i in range(len(id2feat_voc)):
    this_feat_id = str(i)
    this_feat_word = id2feat_voc[this_feat_id]
    prev_feat_id = feature_to_id[this_feat_word]
    assert prev_feat_id not in tracked_prev_feat_ids
    tracked_prev_feat_ids.add(prev_feat_id)
    prev_feat_emb = featureid2emb_voc[prev_feat_id]
    featid2emb_voc[this_feat_id] = prev_feat_emb

featid2emb_file = '../Dataset/{}/train/feature/featureid2embedding.json'.format(dataset_name)
with open(featid2emb_file, 'w') as f:
    print("Write file: {}".format(featid2emb_file))
    json.dump(featid2emb_voc, f)
